Move some functions from daemon.js
